# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "代理订阅售卖系统后端 API。统一信封 {code,message,data}；鉴权 Bearer \u003caccess_token\u003e。",
        "title": "NanoCloud API",
        "contact": {},
        "version": "1.0"
    },
    "host": "api.example.com",
    "basePath": "/api/v1",
    "paths": {
        "/admin/orders/{order_no}/refund": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理端"
                ],
                "summary": "订单退款",
                "parameters": [
                    {
                        "type": "string",
                        "description": "订单号",
                        "name": "order_no",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RefundReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.Body"
                        }
                    }
                }
            }
        },
        "/admin/settings": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理端"
                ],
                "summary": "保存站点设置",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AdminSettingsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.Body"
                        }
                    }
                }
            }
        },
        "/admin/stat/overview": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理端"
                ],
                "summary": "管理端仪表盘",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.AdminOverviewResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/traffic/import": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理端"
                ],
                "summary": "流量导入",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TrafficImportReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.Body"
                        }
                    }
                }
            }
        },
        "/admin/users/{id}/balance": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理端"
                ],
                "summary": "调整用户余额",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AdjustBalanceReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.Body"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.AdjustBalanceReq": {
            "type": "object",
            "required": [
                "amount"
            ],
            "properties": {
                "amount": {
                    "description": "元，可正可负",
                    "type": "number"
                },
                "remark": {
                    "type": "string"
                }
            }
        },
        "model.AdminOverviewResp": {
            "type": "object",
            "properties": {
                "agent_count": {
                    "type": "integer"
                },
                "completed_orders": {
                    "type": "integer"
                },
                "order_count": {
                    "type": "integer"
                },
                "plan_count": {
                    "type": "integer"
                },
                "today_revenue": {
                    "description": "今日实收（元）",
                    "type": "number"
                },
                "total_revenue": {
                    "description": "已完成订单实收合计（元）",
                    "type": "number"
                },
                "user_count": {
                    "type": "integer"
                }
            }
        },
        "model.AdminSettingsReq": {
            "type": "object",
            "required": [
                "key",
                "value"
            ],
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "description": "JSON 字符串",
                    "type": "string"
                }
            }
        },
...
        "model.RefundReq": {
            "type": "object",
            "properties": {
                "remark": {
                    "type": "string"
                }
            }
        },
...
        "model.TrafficImportItem": {
            "type": "object",
            "required": [
                "date",
                "user_id"
            ],
            "properties": {
                "d": {
                    "type": "integer"
                },
                "date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "u": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.TrafficImportReq": {
            "type": "object",
            "required": [
                "items"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/model.TrafficImportItem"
                    }
                }
            }
        },
...
        "resp.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}